pyproject.toml
[project]
name = "teleshell"
version = "0.1.7"
description = "AI-driven Telegram automation and intelligence CLI tool"
authors = [
    {name = "Gemini CLI", email = "gemini-cli@example.com"}
]
dependencies = [
    "telethon>=1.34.0",
    "litellm>=1.81.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.3",
    "click>=8.1.0",
    "google-generativeai>=0.8.6",
    "inquirerpy>=0.3.4"
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.25.0",
    "black>=26.1.0",
    "ruff>=0.15.0",
    "mypy>=1.19.0",
    "respx>=0.22.0",
    "types-pyyaml>=6.0.12.20241230"
]

[project.scripts]
tshell = "teleshell.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
